% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugs.plot.summary.R
\name{bugs.plot}
\alias{bugs.plot}
\alias{bugs.plot.summary}
\alias{bugs.plot.inferences}
\title{Plotting summary information - intended for internal use only}
\usage{
bugs.plot.summary(sims, ...)
}
\arguments{
\item{sims}{an object of class \code{bugs}, see \code{\link{bugs}} for
details}

\item{...}{further arguments to be passed to low-level plot functions}

\item{display.parallel}{display parallel intervals in both halves of the
summary plots; this is a convergence-monitoring tool and is not necessary
once you have approximate convergence (default is \code{FALSE})}
}
\value{
Does not return anything, but prints and plots as side-effects.
}
\description{
Plotting summary information - intended for internal use
}
\details{
\code{bugs.plot.summary} (left hand side of plot) and
\code{bugs.plot.inferences} (right hand side of plot).
}
\seealso{
The main function to be called by the user is \code{plot}, see
\code{\link{plot.bugs}} for details.
}
\keyword{internal}
