% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugs.sims.R
\name{bugs.sims}
\alias{bugs.sims}
\title{MultiBUGS output reader}
\usage{
bugs.sims(parameters.to.save, n.chains, n.iter, n.burnin, n.thin,
  DIC = TRUE)
}
\arguments{
\item{parameters.to.save}{parameters that should be monitored}

\item{n.chains}{number of Markov chains}

\item{n.iter}{number of total iterations (including burn in)}

\item{n.burnin}{length of burn in}

\item{n.thin}{size of thinning parameter}

\item{DIC}{calculation of DIC}
}
\value{
Returns the same values as \code{\link{bugs}}.
}
\description{
Reads simulations from \pkg{MultiBUGS} into , formats them, monitors
convergence, performs convergence checks, and computes medians and quantiles
- intended for internal use.
}
\seealso{
The main function to be called by the user is \code{\link{bugs}}.
}
\keyword{IO}
\keyword{file}
\keyword{internal}
