\name{R2MultiBUGS-package}
\alias{R2MultiBUGS-package}
\alias{R2MultiBUGS}

\docType{package}

\title{Running MultiBUGS from R}

\description{

\pkg{R2MultiBUGS} Call a \pkg{BUGS} model,
summarize inferences and convergence in a table and graph, and save the
simulations in arrays for easy access in \R.  The main command is \code{\link{bugs}}.

}

\details{

The following are sources of information on \pkg{R2MultiBUGS} package:
\tabular{ll}{
  DESCRIPTION file\tab \code{library(help="R2MultiBUGS")}\cr
  \tab \cr
  This file\tab        \code{package?R2MultiBUGS}\cr
  \tab \cr
  Vignette\tab         \code{vignette("R2MultiBUGS")}\cr
  \tab \cr
  Some help files\tab  \code{\link{bugs}}\cr
  \tab                 \code{\link{write.model}}\cr
  \tab                 \code{\link{print.bugs}}\cr
  \tab                 \code{\link{plot.bugs}}\cr
  \tab \cr
  News\tab             \code{file.show(system.file("NEWS", package="R2MultiBUGS"))}\cr
}

}

\keyword{package}
