% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bugs.R
\name{print.bugs}
\alias{print.bugs}
\title{Printing a bugs object}
\usage{
\method{print}{bugs}(x, digits.summary = 1, ...)
}
\arguments{
\item{x}{an object of class `bugs', see \code{\link{bugs}} for details}

\item{digits.summary}{rounding for tabular output on the console (default is
to round to 1 decimal place)}

\item{...}{further arguments to \code{\link{print}}}
}
\description{
Printing a \code{bugs} object
}
\seealso{
\code{\link{bugs}}
}
\keyword{print}
