% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2MultiBUGS-package.R
\docType{package}
\name{R2MultiBUGS-package}
\alias{R2MultiBUGS-package}
\alias{R2MultiBUGS}
\title{Running MultiBUGS from R}
\description{
\pkg{R2MultiBUGS} Call a \pkg{BUGS} model, summarize inferences and
convergence in a table and graph, and save the simulations in arrays for
easy access in \code{R}. The main command is \code{\link{bugs}}.
}
\details{
The following are sources of information on \pkg{R2MultiBUGS} package:
\tabular{ll}{
  DESCRIPTION file\tab \code{library(help="R2MultiBUGS")}\cr
  \tab \cr
  This file\tab        \code{package?R2MultiBUGS}\cr
  \tab \cr
  Some help files\tab  \code{\link{bugs}}\cr
  \tab                 \code{\link{write.model}}\cr
  \tab                 \code{\link{print.bugs}}\cr
  \tab                 \code{\link{plot.bugs}}\cr
}
}
\keyword{package}
