% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.bugs.R
\name{read.bugs}
\alias{read.bugs}
\title{Read output files in CODA format}
\usage{
read.bugs(codafiles, ...)
}
\arguments{
\item{codafiles}{character vector of filenames (e.g. returned from
\code{\link{bugs}} in call such as \code{bugs(....., codaPkg=TRUE, .....)}).
Each of the files contains coda output for one chain produced by
\pkg{MultiBUGS}, the \emph{directory} name of the corresponding file
\file{CODAindex.txt} is extracted from the first element of
\code{codafiles}.}

\item{...}{further arguments to be passed to \code{\link[coda]{read.coda}}}
}
\description{
This function reads Markov Chain Monte Carlo output in the CODA format
produced by \pkg{MultiBUGS} and returns an object of class
\code{\link[coda]{mcmc.list}} for further output analysis using the
\pkg{coda} package.
}
\seealso{
\code{\link{bugs}}, \code{\link[coda]{read.coda}},
\code{\link[coda]{mcmc.list}}
}
\keyword{IO}
\keyword{file}
