\name{schools}
\alias{schools}
\docType{data}
\title{8 schools analysis}
\description{8 schools analysis}
\usage{data(schools)}
\format{
  A data frame with 8 observations on the following 3 variables.
  \describe{
    \item{school}{See Source.}
    \item{estimate}{See Source.}
    \item{sd}{See Source.}
  }
}
\source{
Rubin, D.B. (1981): Estimation in Parallel Randomized Experiments.
\emph{Journal of Educational Statistics} 6(4), 377-400.

Section 5.5 of
Gelman, A., Carlin, J.B., Stern, H.S., Rubin, D.B. (2003): 
\emph{Bayesian Data Analysis}, 2nd edition, CRC Press.
}
\keyword{datasets}
